ceph: integration: upgrade: disable dashboard
Disable the dashboard in the upgrade test because upgrading Ceph from
Mimic to Nautilu with Rook v1.0 can stall for a LONG time when updating
the mgr module. We are pretty sure there is very little risk that the
upgrade test will expose problems with the dashboard module that aren't
caught by other integration tests.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>