CSI: create socket file outside of kubelet plugin directory
`/var/lib/kubelet/plugins` directory is used for
plugin discovery and registration, currently we are creating
the directory with the driver name and placing the socket file
inside the directory, this is causing the spam logs in
kubelet. To avoid this we need to create the directory
outside of kubelet plugin path and place the socket
files inside it.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>