edgeFs: Fix logic of error handling in the finalizer method
If obj is not successfully cast into an e dgefsv1.Cluster, ok will be false, resulting in the .Update(cluster) operation not running, and the declared err variable retaining its default nil value. Subsequently, a check against err is performed, which will evaluate to false, resulting in logging the finalizer as removed when it in fact was not.

Signed-off-by: Nizamudeen <nia@redhat.com>