ceph: allow for empty all annotations
When the annotations are generated for a ceph daemon,
the all annotations are merged with the component annotations.
If there are no all annotations, the component annotation
calculation was crashing. Now we check for the case where
the all annotations are not defined.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 16b2058432a864f329cde6d0c6862da7bde8f7b8)