ceph: remove owner reference from mon PVC
When the cluster CR is deleted, all the resources are also deleted.
To prevent accidental removal of critical data, we don't want to remove
PVCs behind MONs or OSDs automatically. The PVCs behind OSDs already do not
have owner references for this reason. Now this change removes the owner
reference from the MONs as well.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>