ceph: Less generic messages for the cluster Progressing Condition
This commit will changes the message for the Progressing condition to a more generic one.

When the initial reconciliation takes place, for Progressing
first the message will be Cluster is checking for mon health,
once that is done then Cluster starts running mgr,
then it'll be Cluster is checking for osd health.
when everything is done Cluster created succesfully.

If the operator is restarted the message will be Cluster is checking for updates if needed everytime.

Signed-off-by: Nizamudeen <nia@redhat.com>