ceph: extend cluster delete retry window to 15 minutes
Because the ClusterController is event-driven and single-threaded,
we cannot currently block cluster deletion permanently until PVs
are deleted for fear of blocking other users in a multitenant
environment. However, the current cluster delete retry window of
30 seconds is by far our shortest retry window, though it is our
most destructive action. This patch extends the delete retry
window to 15 minutes, which is more in line with our other
operations.

Signed-off-by: Elise Gafford <egafford@redhat.com>