stalebot: add keepalive label to exempt issues
Some issues should not be closed by the stale bot, whether because an
issue is taking a long time to design, implement, or if we need an
owner for the issue, but we know we need the feature or fix.

The keepalive label will cause the bot to skip closing stale issues.

This label should be used sparingly

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>