ceph: Adds design document for admission controller support in rook
Adding admission controllers to rook would help in performing custom CRD validations to rook resources. It ensures that the core specifications that reach rook remain valid despite any kind of changes. In case any invalid changes are found, The kubernetes api server would reject the request and  an error would be sent back to the user.

Signed-off-by: vbnrh <vbadrina@redhat.com>