ceph: use controller runtime logic to watch CephCluster
This is a follow-up on https://github.com/rook/rook/pull/5160 which uses
the controller runtime paradigm approach of watchers. So we now watch
for a CephCluster CR, if the spec image changed we trigger an upgrade.

The downside of that approach is that there is no control over the
workflow and this could allow version rollback.
We could use DiffImageSpecAndClusterRunningVersion but that requires
having a client and passing more to the predicate function. It's not a
hard blocker though, this can be revisited later.

Signed-off-by: Sébastien Han <seb@redhat.com>