ceph: helm add missing subresources to CRD
The subresources field was missing, this was missed in
https://github.com/rook/rook/pull/4964.

The .metadata.generation field is updated if and only if the value at the .spec subpath changes.
Additionally, if the spec does not change, .metadata.generation is not updated.
This is a must-have for the controller-runtime work, if we don't have
this, the generation field will always be incremented, resulting in a
endless reconcile loop.

Closes: https://github.com/rook/rook/issues/5173
Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 0e42222c50f03f5407ea76633b6d0cf93675d942)