ceph: update values.yaml
helm v2.5.1 is currently used, but it is old and unable to use template command.
However, upgrading the helm version from v2.5.1 to v2.16.5 causes lint error.
`helm lint --strict` does not allow to use non-existing values in values.yaml.

Therefore, this commit explicitly declare the values by that which is
evaluated as false in a pipeline.

Signed-off-by: binoue <banji-inoue@cybozu.co.jp>