ceph: allow toolbox script to skip watching mons
The toolbox script is used for configuring the toolbox. If the
mons are to fail over while the toolbox is running, the new
list of mons needs to be updated. However, if run temporarily
as a one-time job, the script only needs to initialize the
ceph.conf and then exit.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit c024f01d9a7f74aa5b561a1b0925e5c443c48bdc)