ceph: metadatadevice to not be added to osd data device list
When the metadataDevice is specified, it was a regression that it was
being interpreted also as a data device. ceph-volume was failing
when the metadataDevice was added as both a data and metadata device.
So we ensure it is not added to the list of data devices.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 99dc363ab6e0b9386a5a33726fe68a482683f0fc)