ceph: enable pg_autoscaler module in default examples
The pg_autoscaler is enabled by default in octopus. In nautilus
it was a new feature and was not enabled by default for that reason.
Now that the pg_autoscaler is more stable and is essentially
the same as the autoscaler in octopus, we can enable it by default.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit d9ee447549022cf465f7e1624514fbf8e6baba42)