ceph: do not fail on deactivate
If we fail to release an LVM device, we should not fail the deployment.
We just print the error and let the backend handle the rest. Some
CSI backend will force detach the device anyway.
This prevents the deployment to get stuck in terminating.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 65aae84bdd1c1975b5fde1c16328be185712fe66)