ceph: allow csi driver version to be undetected
If ROOK_CSI_ALLOW_UNSUPPORTED_VERSION=false and the csi canary image
is set to quay.io/cephcsi/cephcsi:canary, the csi driver fails
to load as expected since it is not supported. But if the
flag is set to true, it is expected to allow the canary image
which isn't versioned. The supported version check was requiring
that the version be detected in the image.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit b375e81c37ae2f1a87333e846a824492600c02ce)