ceph: skip ceph-volume inventory for PVs
For multipath devices, ceph-volume inventory is failing
to detect that the device is available, complaining about
insufficient space. Instead of relying on c-v inventory
to determine if the device is available, just call
c-v raw list to determine if there is already an osd
configured on the pv. Otherwise, just attempt to configure
the OSD.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>